moved maat refs populate logic out of billing package
