+++ maat-scheduled-tasks/src/main/java/uk/gov/justice/laa/maat/scheduled/tasks/repository/MaatReferenceRepository.java
@@ -1,11 +1,11 @@
-package uk.gov.justice.laa.maat.scheduled.tasks.billing.repository;
+package uk.gov.justice.laa.maat.scheduled.tasks.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
-import uk.gov.justice.laa.maat.scheduled.tasks.billing.entity.MaatReferenceEntity;
-import uk.gov.justice.laa.maat.scheduled.tasks.billing.entity.MaatReferenceId;
+import uk.gov.justice.laa.maat.scheduled.tasks.entity.MaatReferenceEntity;
+import uk.gov.justice.laa.maat.scheduled.tasks.entity.MaatReferenceId;
 
 @Repository
 public interface MaatReferenceRepository extends
