+++ maat-scheduled-tasks/src/main/java/uk/gov/justice/laa/maat/scheduled/tasks/service/MaatReferenceService.java
@@ -1,10 +1,10 @@
-package uk.gov.justice.laa.maat.scheduled.tasks.billing.service;
+package uk.gov.justice.laa.maat.scheduled.tasks.service;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import uk.gov.justice.laa.maat.scheduled.tasks.billing.repository.MaatReferenceRepository;
+import uk.gov.justice.laa.maat.scheduled.tasks.repository.MaatReferenceRepository;
 import uk.gov.justice.laa.maat.scheduled.tasks.exception.MAATScheduledTasksException;
 
 @Slf4j
